Add optional encryption of recovery codes.
The optional encryption uses cryptography.fernet.

Refactor the recovery codes code to put it all in one file, and add an overridable class - MfRecoveryCodesUtil that controls
creating, checking, encrypting, decrypting recovery codes.

closes #693
