+++ docs/configuration.rst
@@ -1554,6 +1554,25 @@ Recovery Codes
 
     Default: ``"security/mf_recovery.html"``
 
+.. py:data:: SECURITY_MULTI_FACTOR_RECOVERY_CODES_KEYS
+
+    A list of keys used to encrypt the recovery codes at rest (i.e. in the database).
+    The default implementation uses cryptography.fernet (https://cryptography.io/en/latest/fernet/#cryptography.fernet.Fernet)
+    - so the keys should be generated by::
+
+        from cryptography.fernet import Fernet
+        key = Fernet.generate_key()
+
+    Multiple keys can be configured allowing for key rotation.
+
+    Default: ``None`` - recovery codes will NOT be encrypted on disk
+
+.. py:data:: SECURITY_MULTI_FACTOR_RECOVERY_CODE_TTL
+
+    An integer passed to decrypt specifying the maximum age of the code.
+
+    Default: ``None`` - no TTL will be enforced.
+
 Additional relevant configuration variables:
 
     * :py:data:`SECURITY_FRESHNESS` - Used to protect /mf-recovery-codes.
@@ -1587,8 +1606,8 @@ A list of all URLs and Views:
 * ``SECURITY_RESET_URL``
 * ``SECURITY_CHANGE_URL``
 * ``SECURITY_CONFIRM_URL``
-* :py:data:``SECURITY_MULTI_FACTOR_RECOVERY_CODES_URL``
-* :py:data:``SECURITY_MULTI_FACTOR_RECOVERY_URL``
+* :py:data:`SECURITY_MULTI_FACTOR_RECOVERY_CODES_URL`
+* :py:data:`SECURITY_MULTI_FACTOR_RECOVERY_URL`
 * :py:data:`SECURITY_TWO_FACTOR_SELECT_URL`
 * :py:data:`SECURITY_TWO_FACTOR_SETUP_URL`
 * :py:data:`SECURITY_TWO_FACTOR_TOKEN_VALIDATION_URL`
