+++ flask_security/datastore.py
@@ -551,11 +551,6 @@ class UserDatastore:
         """Set MF recovery codes into user record.
         Any existing codes will be erased.
 
-        .. danger::
-           Be aware that whatever `passwords` are passed in will
-           be stored directly in the DB. Do NOT pass in a plaintext passwords!
-           Best practice is to pass in ``hash_password(plaintext_password)``.
-
         .. versionadded: 5.0.0
         """
         user.mf_recovery_codes = rcs
@@ -565,7 +560,7 @@ class UserDatastore:
         codes = getattr(user, "mf_recovery_codes", [])
         return codes if codes else []
 
-    def mf_delete_recovery_code(self, user: "User", code_to_delete: str) -> bool:
+    def mf_delete_recovery_code(self, user: "User", idx: int) -> bool:
         """Delete a single recovery code.
         Recovery codes are single-use - so delete after using!
 
@@ -576,10 +571,10 @@ class UserDatastore:
         if not user.mf_recovery_codes:
             return False
         try:
-            user.mf_recovery_codes.remove(code_to_delete)
+            user.mf_recovery_codes.pop(idx)
             self.put(user)
             return True
-        except ValueError:
+        except IndexError:
             return False
 
     def us_get_totp_secrets(self, user: "User") -> t.Dict[str, str]:
